move year svelte page into different dictionary
